Form ammendment and login Form Code

+++ Project/LoginHandler.cs
@@ -11,7 +11,7 @@ namespace Project
     {
         private static string filename = @"accounts.txt";
 
-        public static bool Login(string username, string password)
+        public bool Login(string username, string password)
         {
             bool flag = false;
             if (File.Exists(filename) != true)
@@ -33,16 +33,45 @@ namespace Project
             return flag;
         }
 
-        public static void CreateAccount(string username, string password)
+        public string CreateAccount(string username, string password, string conPass)
         {
+            bool Exists = false;
+            string Message = "";
             if (File.Exists(filename) != true)
             {
-                File.CreateText(filename).Close();
+                File.CreateText(filename).Close();   
             }
-            FileStream fs = new FileStream(filename, FileMode.Append);
-            StreamWriter writer = new StreamWriter(fs);
-            writer.WriteLine(username + "," + password);
-            writer.Close();
+            using (StreamReader reader = new StreamReader(filename))
+            {
+                string line;
+                while ((line = reader.ReadLine()) != null)
+                {
+                    string[] details = line.Split(',');
+                    if ((username == details[0]) && (password == details[1]))
+                    {
+                        Exists = true;
+                    }
+                }
+            }
+            if (Exists == true)
+            {
+                Message = "The username already exists";
+            }
+            else if ((Exists == false) && (password != conPass))
+            {
+                Message = "The password and the confirm password does not match.";
+            }
+            else if ((Exists == false) && (password == conPass))
+            {
+                FileStream fs = new FileStream(filename, FileMode.Append);
+                StreamWriter writer = new StreamWriter(fs);
+                writer.WriteLine(username + "," + password);
+                writer.Close();
+
+                Message = "Account successfully created";
+            }
+
+            return Message;
         }
     }
 }
